Module mems_office::db::structs

Structs

E01

Represents the E01 form in the database

Furniture

Represents the Furniture form in the database

Instrument

Leave

Represents the Leave form in the database

MM04

Represents the MM04 form in the database

R1

Represents the R1 form in the database

Receivers

SS01

Represents the SS01 form in the database

SS01Orders

Represents the list_orders in SS01 form in the database

SS04

Represents the SS04 form in the database

SS04Orders

This entity represents the list_orders of SS04 table

Seeking

It is used to fetch the seeking, pending or previous forms from the database

Student

UserDb

Represents the user table in the database

Users

Enums

Degree

Represents the academic program of the Student users which are:

Department

Represents the possible departments in the college which are:

Desig

Represents the designation of the users that are going to use the website which are:

State

Represents the possible states of the form which are: Pending, if the form is not yet accepted/rejected by the receiver; Accepted, if the form is accepted by the receiver and Rejected if the form is rejected by the receiver