Module mems_office::lists

Modules

fetch_forms 🔒

Structs

create_students

Functions

get_receiver

This function is used to fetch all the possible receivers of the forms, so the form sender can choose who to send

get_students

list_config

Configures the routes for the list module in the Actix web server.