Trait mems_office::view::forms::FormTrait

pub trait FormTrait: Serialize {
    // Required methods
    async fn process(
        &self,
        pool: &PgPool,
        id: i32
    ) -> Result<(), Box<dyn Error>>;
    async fn get_identifier(
        &self,
        pool: &PgPool
    ) -> Result<Identifier, Box<dyn Error>>;
    async fn pg_insert(&self, pool: &PgPool) -> Result<i32, Box<dyn Error>>;

    // Provided methods
    async fn send_form_ws(
        &self,
        srv: &mut ChatServer,
        pool: &PgPool
    ) -> Result<(), Box<dyn Error>> { ... }
    async fn send_recv_update(
        &self,
        pool: &PgPool,
        id: i32,
        form_name: &str
    ) -> Result<(), Box<dyn Error>> { ... }
}

Required Methods

async fn process(&self, pool: &PgPool, id: i32) -> Result<(), Box<dyn Error>>

This function is used to process the form. It takes in the pool object and the id of the form and returns a Result object.

async fn get_identifier( &self, pool: &PgPool ) -> Result<Identifier, Box<dyn Error>>

This function is used to get the Identifier object corresponding to the form. It takes in the pool object and returns a Result object.

async fn pg_insert(&self, pool: &PgPool) -> Result<i32, Box<dyn Error>>

This function is used to insert the form into the database. It takes in the pool object and returns a Result object.

Provided Methods

async fn send_form_ws( &self, srv: &mut ChatServer, pool: &PgPool ) -> Result<(), Box<dyn Error>>

This function is used to send the form to the WebSocket server. It takes in the server object and the pool object and returns a Result object.

async fn send_recv_update( &self, pool: &PgPool, id: i32, form_name: &str ) -> Result<(), Box<dyn Error>>

This function is used to update the sender and receiver of the form. It takes in the pool object, the id of the form and the form_name and returns a Result object.

Object Safety

This trait is not object safe.